replaces old kafka partition impls w/ refactored variants
