+++ pkg/kafka/partition/committer_test.go
@@ -36,7 +36,7 @@ func TestPartitionCommitter(t *testing.T) {
 	reg := prometheus.NewRegistry()
 	partitionID := int32(1)
 	consumerGroup := "test-consumer-group"
-	reader := NewRefactoredReader(
+	reader := newReader(
 		client,
 		kafkaCfg.Topic,
 		partitionID,
@@ -44,7 +44,7 @@ func TestPartitionCommitter(t *testing.T) {
 		logger,
 		reg,
 	)
-	committer := newRefactoredCommitter(reader, kafkaCfg.ConsumerGroupOffsetCommitInterval, logger, reg)
+	committer := newCommitter(reader, kafkaCfg.ConsumerGroupOffsetCommitInterval, logger, reg)
 
 	// Test committing an offset
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
